Code cleanup.
git-svn-id: https://src.develer.com/svnoss/bertos/trunk@1033 38d2e660-2303-0410-9eaa-f027e97ec537

+++ mware/byteorder.h
@@ -69,8 +69,8 @@ INLINE uint32_t swab32(uint32_t x)
  */
 INLINE uint64_t swab64(uint64_t x)
 {
-	return ( (uint64_t)swab32((x >> 32))
-		| (uint64_t)swab32(x & (uint32_t)0xFFFFFFFFUL) << 32);
+	return (uint64_t)swab32(x >> 32)
+		| ((uint64_t)swab32(x & 0xFFFFFFFFUL) << 32);
 }
 
 /**
